requirements.txt
langchain
python-dotenv
ipykernel
langchain-community

wikipedia
langchain-text-splitters

sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
duckdb
pandas
openai
langchain-groq
langchain_core
fastapi
beautifulsoup4
uvicorn
langchain
python-dotenv
ipykernel
langchain-community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
langchain-openai
chromadb
lxml
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
duckdb
pandas
openai
langchain-groq
langchain_core

fastapi
duckduckgo_search
pymupdf
arxiv
wikipedia
mysql-connector-python
SQLAlchemy
validators
youtube_transcript_api
unstructured
pytube
numexpr
huggingface_hub
beautifulsoup4
pymysql
kdbai_client
uvicorn
Gradio